feat: add base routing and separation into components
