+++ src/App.scss
@@ -3,49 +3,3 @@
   margin: 0 auto;
   text-align: center;
 }
-
-.result {
-  background-color: aqua;
-  width: 100%;
-
-  &__loading {
-    text-align: center;
-    min-height: calc(100vh - 88px - 40px);
-  }
-}
-
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-  transition: filter 300ms;
-  animation-name: logo-spin;
-  animation-duration: 2s;
-  animation-iteration-count: infinite;
-  animation-timing-function: linear;
-}
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
-
-@keyframes logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.error-block {
-  background-color: black;
-  height: 48px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  &__button {
-    background-color: red;
-    height: 40px;
-    width: 100px;
-  }
-}
