+++ src/main.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
 import './index.scss';
-import ErrorBoundary from './pages/error-page.tsx';
+import ErrorPage from './pages/ErrorPage/ErrorPage.tsx';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <ErrorBoundary>
+    <ErrorPage>
       <App />
-    </ErrorBoundary>
+    </ErrorPage>
   </React.StrictMode>
 );
