+++ src/pages/CardPage/CardPage.scss
@@ -1,0 +1,3 @@
+.card-page {
+  width: 50%;
+}
