+++ src/pages/CardPage/CardPage.tsx
@@ -1,0 +1,6 @@
+import './CardPage.scss';
+const CardPage = () => {
+  return <div className="card-page">Page Work!</div>;
+};
+
+export default CardPage;
