+++ src/pages/MainPage/MainPage.scss
@@ -1,0 +1,22 @@
+.main-page {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.result {
+  background-color: aqua;
+}
+
+.error-block {
+  background-color: black;
+  height: 48px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  &__button {
+    background-color: red;
+    height: 40px;
+    width: 100px;
+  }
+}
